Divide exercises to practice and scored (can be both too).
- Until now, the exercise had only a single boolean flag "active". I am
not even sure, how it was intended at first place. Now "active" means
"prepared and visible to everybody".
- Exercise's boolean flag "scored" means that it is possible to submit
it only once and a student will get points according to their
performance.
- Exercise's boolean flag "practice" means that the exercise is enabled
for training and users can do it multiple times.
- Exercises on home page are divided into "Scored" and "For training".
- Exercise can be both "scored" and "practice". This way it will appear
in "Scored" section and "For training" section too. If user selects the
link in "For training" section, the exercise can not be submitted, only
evaluated and vice versa for "Scored" section.
- Admins can see a summary table for students in their Home.
- Users has new attribute "Enrolled". This is intended entirely for
admins to select which students can be seen in summary table.